Added local .txt file support
Adds the ability to check a local .txt file for changes when song requests are disabled or not connected.

Added commands:
- !ytp togglelocal - enable or disable local file functionality
- !ytp localpath [./path/to/file.txt] - change the path to the .txt file. MUST end in .txt ( command checks for this )

Modified commands:
- !currentsong - reads from local file if song requests aren't available, responds with "No song is currently playing" if neither are available. As a result this command has moved above the check for the connected player client in order to allow it to run.

A note about the 3 if statements added at the top for local functionality: if this is added to the release version, these could be simplified if the variables were included in the stock database and could use the same format as the rest of the variables such as playlistDJname. In the current version the more complex if statements allow for these to be missing.