Added !cooltype [command] [type] command
!cooltype [command] [type]
This command allows you to override PhantomBot's logic for figuring out which type of cooldown to use for a specific command. This is especially useful for scenarios where you have a global cooldown time but you want to change a few commands to have no cooldown, or if you want to enable both global and per-user cooldowns but you want to specify some commands' global cooldown times.

The four types of cooldown "types" are:
A) default - this is the default setting, and allows PhantomBot's logic to figure out what type of cooldown each command follows
B) per-user - this overrides PhantomBot's logic and enforces per-user cooldowns for whatever commands you give this type to
C) global - this overrides PhantomBot's logic and enforces global cooldowns for whatever commands you give this type to
D) none - this overrides PhantomBot's logic and enforces no cooldowns for whatever commands you give this type to

Example Scenarios:
A)
    !toggleglobalcooldown
    !toggleperusercooldown
    !globalcooldowntime 2
    !cooltype !8ball global
    !coolcom !8ball 30

B)
    !toggleglobalcooldown
    !globalcooldowntime 2
    !coolcom !slot none

C)
    !toggleperusercooldown
    !cooltype !8ball global
    !cooltype !slot global
    !globalcooldowntime 10

The main reason I did this was because I noticed that with the way PhantomBot's logic currently works, it is impossible to specify specific commands to have a specific global cooldown time (AKA not the generic !globalcooldowntime) if you have both per-user and global enabled. There are some other scenarios that this !cooltype concept fixes as well, but I won't get into listing all of them since the above case should be enough to merit this change.