Backup for SQLite3 DB
Note that this only is implemented for SQLite3.  There are no plans to include support for
MySQL although IniStore may be supported in the future.  MySQL will not be supported as
it is up to the MySQL administrator of the PhantomBot database to handle backup and
restore operations.  There are various options for providing backup services for MySQL.

**DataStore.java**
- Template function for backing up DB.

**SqliteStore.java**
- Backup for SQLite3 database.

**PhantomBot.java**
- Support to manually backup SQLite3 database
	- Console: !backupdb
	- Chat: !debug !dev !dbbackup

- Support to automatically backup SQLite3 database via botlogin.txt configuration
	- backupsqliteauto=true
	- Optional: backupsqlitehourfrequency=xx hours to wait between backups, 24 if not provided
	- Optional: backupsqlitekeepdays=xx days to keep automatic backups, 5 if not provided