BETA - Detect Clips
There are restrictions on this due to the Twitch API.  Twitch allows us to poll the
clips from the "day" (assuming this is the current day, not sure if it is last 24 hours
or not though).  We poll 100 of the clips.  We then scan over all of the clips for the
newest one.  If this is newer than the last one in the database, then an event will be
sent and a message presented in chat.

Keep in mind that if you have over 100 clips in the "day" period, that the newest one
will only bubble up if it has more views than the last newest clip.  This, unfortunately,
is what Twitch returns and we hope they provide a better endpoint in the future.

This endpoint will be polled every 60 seconds.

**clipHandler.js**
- Handles the event from Twitch and displays a message in chat
- Supports !clipstoggle and !clipsmessage
- !clipsmessage supports (name) and (url) tag
- Disabled by default

**init.js**
- Hook the twitchClip event

**handlers-clipHandler.js**
- Language entries, the usage and output for !clipstoggle and !clipsmessage

**TwitchAPIv5.java**
- Added new method for grabbing the clips from the "day" with a limit of 100 (max)

**TwitchCache.java**
- If !clipstoggle is enabled, will query the clips API every 60 seconds and return an event if applicable

**TwitchClipEvent.java**
- Event definition