- Moved verified check to debug-only-mode
**PhantomBot.java:**
- Moved this console log to debug only mode because users that use the
nightly are already asking about this, once we push it to stable we'll
get a lot more. Plus, for every basic user they do not need to get their
bot's verified, it changes nothing to them. 20 messages for one channel
is more than enough. Plus, Twitch hasn't been accepting new bots for a
few weeks now, so this will lead to more confusion.