- Viewer list cache optimization and Discord stream alerts
**ViewerListCache.java, init.js, permissions.js:**
- Changed the logic of getting viewers from TMI. The core will now send
one event with all joins and parts and new users list to a new event
called `IrcChannelUsersUpdate` every 10 minutes which will run on a new
thread in Rhino to update the lists. This can take 3 minutes to run with
36,000 viewers but doesn't slow down events like the old method did. The
old method would queue all events which would cause the bot to slow
down.

**streamHandler.js:**
- You can now ping people with a role on the live messages.