- Switched back to unlimited message parsing threads
**TwitchWSIRC.java:**
- After a few weeks of trying to find the root cause of the bot slowdown
in larger channels, I pinned down the issue to the thread pool running out
of threads when Twitch sends bursts of messages. The eventbus can take
some time to queue events in async (600ms at times), which could cause all threads in the
pool to be active, thus slowing down message processing. This was tested
with a custom IRC server to make sure this was the real issue.