Implemented Twitch UserID tracking and automatic name change handling Completes trello card https://trello.com/c/hl95xq7s
*** tv.phantombot.event.twitch ***
Added package tv.phantombot.event.twitch.user
Added class tv.phantombot.event.twitch.user.TwitchUserNameChanged with properties userID, oldName, and newName

*** UsernameCache ***
Added fields int intervalLookupMS, ScheduledThreadPoolExecutor tpExec, and Queue lookupQueue
Changed type of field cache to ConcurrentMap
Replaced deprecated call to Maps.newHashMap with diamond inference
Added overloads to method addUser
Replaced all calls to cache.put and lookupUserData with addUser
Added private method checkUserNameChanged
Changed addUser to add to the cache if all data is available
Changed addUser to enqueue usernames in lookupQueue if all data is not available

*** init.js ***
Added bind for twitchUserNameChanged

*** nameConverter.js ***
Moved name change execution logic to seperate function
Changed name change execution logic to use a transaction
Counted updating one or more quotes as +1 to number of changed tables
Updated !namechange to call the function
Added bind for twitchUserNameChanged

*** permissions.js ***
Added call to $.username.addUser to binds for ircChannelMessage, ircChannelJoin, and ircChannelUsersUpdate

*******************************

This update implements tracking of Twitch UserIDs. When a user enters the channel or gets resolved,
their information is added to the UsernameCache using IRCv3 tags, as before, if available.

If it is not available, the user is now queued to have their information looked up. This is
implemented as being done at a rate of 1 user per second, to prepare for helix rate limits.
The downside of this change is that if a user is resolved before they have been looked up, their
username will be returned instead of doing an instant lookup.

When addUser successfully detects valid data and adds the user to the cache, it calls a method
that looks up the UserID in the new userids table. If the UserID (table key) is found, it checks
the stored username (table value) against the username being cached. If they do not match, the
TwitchUserNameChangedEvent is posted to the EventBus. The stored username is inserted if it doesn't
exist or updated if it has changed.

init.js handles the event as a forced hook call, so that disabled scripts handling their own
name changes can update their tables in case they later get re-enabled.

The new overloads for addUser allow data to be inserted from existing data, or IRCv3 tags. addUser
now considers userID invalid, and therefore enqueues a user for lookup, if the userID is 0.

The changing of cache to a ConcurrentMap type is required for thread safety.

The use of a ScheduledThreadPoolExecutor is preferable to an infinite loop with Thread.sleep or a Timer.