Added recording the top cheer
Added recording the top cheer to database and text file. This is just a simple addition for people who use the text files to display information on stream. Something could be added to the message in the future, I haven't messed around with the language files yet, so idk. The saved values are in phantombot_streamInfo where I also found latestSub, so that should be fine.