Fake purge and color message filters regex update
userName: /me Test message #1
[CHAT] .timeout userName 5 Using colored text. (Automated by botName)
[CHAT] @userName, you were timed out for using colored text. (warning)
userName: /ME Test message #2
[CHAT] .timeout userName 600 Using colored text. (Automated by botName)

- - - - - - - - -

At the moment regex `/^<message \w+>|^<\w+ deleted>/i` catches only messages starting with `<message` and/or containing `deleted>`. Examples: `<message deleted>` as it is, `<mEsSaGe annihilated>` and `<My message was unfairly deleted> WTF?`. The last two are definitely not fake purge, but a joke (or attempts to make a joke) or wonder.

Also Twitch added a different purge message for its mobile applications, and it looks like this: `<Message deleted.>`.

My suggestion is to improve regex this way:
`/(^<message\s+deleted\.?>$)/i`.

Transcription:
- `^<`… – `<`… must be at the beginning of a message;
- `\s+` – one or more spaces between the `message` and `deleted`;
- `\.?` – null or one dot `.` (for `<message deleted>` in a browser and `<Message deleted.>` in the mobile app);
- …`>$` – …`>` must be at the end of a message.

After such improvement the script catches official Twitch messages `<message deleted>` and `<Message deleted.>` in any case (upper or lower), but does not catches messages like before (examples: "`<message unreasonably deleted>`" or `<message deleted> why???` or even "`<sh*t deleted> FFS!!! Kappa`") that actually are not fake purges but jokes or insults (which are not a duty of fake purge filter but a duty of a moderators).