fixed NullPointerException
Properies implement the Map interface, which provides a put() method.
But if you use put, you can't get the value by getProperty(). I didn't
know that. I changed setting the property correctly. Than I thought
about getting the value savely in the PhantomBot class by using the
ConfigurationManager.getBoolean() method. That would result in never
getting an error, but makes it harder to debug, if something is wrong
with setting the "newSetup" property in future.

** ConfigurationManager.java **
- replaced put by setProperty