Performance improvement for Data Render Service.
Currently, the Times API is computationally intensive because it makes one DB lookup to get the list of keys in the `time` table, and then another three `time` lookups for _each_ key to get its value. These database calls are the performance bottleneck. (The other parts of the DRS handler do something similar, but the `time` table is largest by orders of magnitude -- at least in my case, and it seems reasonable that that would be common.)

This change adds a `GetKeyValuesList()` method to the `DataStore` interface, so that we can fetch the keys and values in a single DB lookup. On my machine (SQLite DB, 153k rows in the `phantombot_time` table) this reduces the total execution time from 40 minutes to 15 seconds.

**javascript-source/handlers/dataServiceHandler.js**
- For the Times data, call `GetKeyValueList()`, and iterate over the array of `KeyValue`s, instead of calling `GetKeyList()`.

**javascript-source/systems/ranksSystem.js**
- Add an optional `time` parameter to `hasRank()` and `getRank()` so that we can skip the DB lookup if we've already done it, but still use the rest of the ranks system logic. The behavior is unchanged if `time` is unset, so other call sites can be left as-is.

**source/com/gmt2001/datastore/Datastore.java**
- Add a `GetKeyValueList()` method, with the same arguments as `GetKeyList()`, to retrieve keys and values simultaneously.

**source/com/gmt2001/datastore/KeyValue.java**
- New immutable data type for String key-value pairs. An array of `KeyValue`s is the return type from `GetKeyValueList()`.

**source/com/gmt2001/datastore/H2Store.java**
**source/com/gmt2001/datastore/IniStore.java**
**source/com/gmt2001/datastore/MySQLStore.java**
**source/com/gmt2001/datastore/SqliteStore.java**
- Implementations of `GetKeyValueList()` for each database type.