- Fix for bot losing moderation powers
Since the "user-type" tag in deprecated and Twitch wants us to reply on badges, a user can sometimes lose its badge for no reason from one message to another. So, this could possibly make the bot lose its moderator powers, even though the bot is still a mod, so checking the "mod" tag before removing a user's moderator permissions is the only way to fix this, an admin or staff member could lose moderation powers from the bot for a second if they are not a channel moderator, so the bot would try and time them out if a moderation filter is triggered. This fix is only to prevent the bot from losing moderator powers.