Fix line 42. IF logic is not triggered currently.
Line 42 originally reads as:
if (streamerGame == null || streamerGame.length === 0) {

This logic does not activate when using !shoutout (streamer) against a Twitch user that has not streamed. It is a less-common use-case, but the logic does exist and having that fallback available in case someone is shouted out that does not actually stream is an appreciated feature. Thanks to the dev for including it to begin with.

The proposed new line 42 reads as:
if (streamerGame == null || streamerGame.length === 0 || streamerGame == "") {

My comfort zone is solely with Powershell, but I know my way around enough to make some small tweaks.
From what I have read when I started looking into this issue, there are 6 "falsy" values used in javascript. (false, 0, "", null, undefined, NaN). The original logic accounts for the string length of streamerGame being 0 OR streamerGame being null (or by association 'undefined'). My proposed change adds a loose comparison for the other "falsy" values by loosely comparing (==) streamerGame against the empty quotes (""). In the loose comparison, "" is equitable to 'false' as well as '0'. I was able to test this change on my own twitch channel and it corrected the problem for me.

Before the change, using !shoutout against a "streamer" (aka. Twitch User) that has not streamed would result in the "corecommands.shoutout.offline" response being triggered and sent in chat by PhantomBot. The logic on the original line 42 did not trigger due to whatever value was captured in variable "streamerGame". Since the comparisons are evaluated as 'this OR that', I added an addition comparison. With the added comparison (streamerGame == ""), it is now using the appropriate "corecommands.shoutout.offline" response when shouting out a streamer/Twitch User that has not streamed.