Optional offline warning msg for raidHandler
Adds raidOutOfflineMessage support, including a default English message.  If PB uses !raid to a channel that is not streaming, a warning will appear in the chat.  It also stores your custom raidOutOfflineMessage in the db.

Also made a slight grammar clarification (line 152) and moved the .raid command earlier so the sequence of messages reads better in chat.