Change declaration order to allow panel to open
Per this thread:
https://discord.com/channels/107910097937682432/1132292565551566858
 Panel would not load, claiming
`helpers.getBranch is not a function`
I assume because it was not yet a function in the order of declarations. I moved helpers.canEmbedTwitch and helpers.CANT... down below where it was declared and it works now.

Disclaimer:  I am using http, not https.  It is possible that someone with https would get past this error, though I doubt it, because it's in the initial declarations.  Maybe those who aren't seeing this error are lucky enough to have the older js cached.