Update custom.js to allow quick toggle of disabled/hidden right on the page
The Disabled and Hidden indicator icons are now clickable.  Switching from a normal stream to a raid train to a different raid train, etc, can be a bit tedious, with several clicks and some scrolling in order to enable/disable the relevant commands.  I have long wished that the indicator icons could be clickable, and finally did it.